+++ app/Http/Controllers/API/UserController.php
@@ -15,6 +15,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        $this->authorize('ADMIN');
+        
         return User::all();
     }
 
@@ -26,15 +28,18 @@ class UserController extends Controller
      */
     public function store(Request $req)
     {
+        $this->authorize('ADMIN');
 
         $fields = $req->validate([
             'username' => 'required|string|unique:users,username',
             'password' => 'required|string|confirmed',
+            'role_id' => 'string',
         ]);
-
+        
         $user = User::create([
             'username' => $fields['username'],
             'password' => bcrypt($fields['password']),
+            'role_id' => $fields['role_id'] ?? 3,
         ]);
 
         $token = $user->createToken('$3cr37')->plainTextToken;
@@ -55,6 +60,8 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        $this->authorize('ADMIN');
+
         return $user;
     }
 
@@ -67,6 +74,8 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        $this->authorize('ADMIN');
+
         $request->validate([
             'username' => 'required|string|unique:users,username',
             'password' => 'required|string',
@@ -85,6 +94,8 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('ADMIN');
+
         $user->delete();
 
         return response([
